Add multithreading (#47)
